Implement a not-overly-DRY production setup... Recommendations welcome!
