+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "torontojs.com",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "The website for TorontoJS",
   "main": "index.js",
   "scripts": {
